+++ includes/publisher/class-publisher-form-editing-experience.php
@@ -380,6 +380,16 @@ final class FormEditingExperience {
 					}
 				} else {
 					$existing = get_post_meta( $post_id, sanitize_text_field( $slug ), true );
+					// Clean up empty values already saved in db.
+					if ( $existing === '' || $existing === [] ) {
+						$deleted = delete_post_meta( $post_id, sanitize_text_field( $slug ) );
+						if ( ! $deleted ) {
+							/* translators: %s: atlas content modeler field slug */
+							$this->error_save_post = sprintf( __( 'There was an error deleting the %s field data.', 'atlas-content-modeler' ), $slug );
+						}
+						continue;
+					}
+
 					if ( empty( $existing ) ) {
 						continue;
 					}
@@ -425,6 +435,12 @@ final class FormEditingExperience {
 
 		foreach ( $posted_values as $key => $value ) {
 			$key = sanitize_text_field( $key );
+
+			// Do not save empty values.
+			if ( empty( $value ) ) {
+				continue;
+			}
+
 			/**
 			 * Check if an existing value matches the submitted value
 			 * and short-circuit the loop. Otherwise `update_post_meta`