+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Text fields can now use “title” as their API identifier if “use this field as the entry title” is ticked.
 - Field validation now returns translatable errors.
 
+### Fixed
+- Issue where adding a new repeating field to an existing model schema could break GraphQL queries under certain conditions.
+- Empty field values are no longer saved to the database.
+
 ## 0.17.0 - 2022-05-05
 ### Added
 - Mutations support to create, update and delete entries via GraphQL. All fields are supported except media and relationships for now. Find examples at https://github.com/wpengine/atlas-content-modeler/blob/main/docs/mutations/index.md.
@@ -27,6 +31,7 @@
 ### Fixed
 - Title values are now saved to the wp_posts table as expected, instead of being exposed with WP filters. This fixes a few things, such as queries that search the post title field.
 - Post slugs are now generated from the post title value, like they are for post types built into WordPress.
+- Taxonomy terms that already exist are skipped when importing a blueprint.
 
 ## 0.16.0 - 2022-04-06
 ### Added